Add Ambient Display QS Tile
Fix Ambient Display tile

Bring in @Altaf-Mahdi code for detecting whether ambient display is on or not
from our lollipop branch

https://github.com/DirtyUnicorns/android_frameworks_base/blob/lollipop/packages/SystemUI/src/com/android/systemui/qs/tiles/AmbientDisplayTile.java

Previous patch from @Noeljunior was not working as it should and so tile was off
although the actual switch was on in Settings. A few toggles of the switch and it
was set to normal but that's janky and users have come to expect better from us :-)

Change-Id: I791400078ea8e3fc1c07134c6a6a76979d81bc64

Point Amibent display tile longpress action to it's own settings [1/2]

http://jira.dirtyunicorns.com/browse/MAR-49

Change-Id: I6cf9e7daaeaff859f7b866a00ac562d39cee568d

Tiles: Remove unused observer in AmbientDisplayTile

The object in question is instantiated but setListening is
never invoked on it. The target URI is already observed making
this object redundant.

Change-Id: I0eef2c36e4138be852af0741201016dd3e753c96